feat(snapshot): move snapshot storage to new file
