+++ src/server/detail/save_stages_controller.cc
@@ -38,29 +38,8 @@ namespace fs = std::filesystem;
 
 namespace {
 
-const size_t kBucketConnectMs = 2000;
-
-#ifdef __linux__
-const int kRdbWriteFlags = O_CREAT | O_WRONLY | O_TRUNC | O_CLOEXEC | O_DIRECT;
-#endif
-
-constexpr string_view kS3Prefix = "s3://"sv;
-
 bool IsCloudPath(string_view path) {
   return absl::StartsWith(path, kS3Prefix);
-}
-
-// Returns bucket_name, obj_path for an s3 path.
-optional<pair<string, string>> GetBucketPath(string_view path) {
-  string_view clean = absl::StripPrefix(path, kS3Prefix);
-
-  size_t pos = clean.find('/');
-  if (pos == string_view::npos)
-    return nullopt;
-
-  string bucket_name{clean.substr(0, pos)};
-  string obj_path{clean.substr(pos + 1)};
-  return make_pair(move(bucket_name), move(obj_path));
 }
 
 string FormatTs(absl::Time now) {
@@ -94,32 +73,6 @@ void ExtendDfsFilenameWithShard(int shard, string_view extension, fs::path* file
   SetExtension(absl::Dec(shard, absl::kZeroPad4), extension, filename);
 }
 
-// takes ownership over the file.
-class LinuxWriteWrapper : public io::Sink {
- public:
-  LinuxWriteWrapper(fb2::LinuxFile* lf) : lf_(lf) {
-  }
-
-  io::Result<size_t> WriteSome(const iovec* v, uint32_t len) final;
-
-  error_code Close() {
-    return lf_->Close();
-  }
-
- private:
-  unique_ptr<fb2::LinuxFile> lf_;
-  off_t offset_ = 0;
-};
-
-io::Result<size_t> LinuxWriteWrapper::WriteSome(const iovec* v, uint32_t len) {
-  io::Result<size_t> res = lf_->WriteSome(v, len, offset_, 0);
-  if (res) {
-    offset_ += *res;
-  }
-
-  return res;
-}
-
 }  // namespace
 
 GenericError ValidateFilename(const fs::path& filename, bool new_version) {
@@ -149,65 +102,6 @@ GenericError ValidateFilename(const fs::path& filename, bool new_version) {
                          "\" for SAVE with type RDB")};
   }
   return {};
-}
-
-FileSnapshotStorage::FileSnapshotStorage(FiberQueueThreadPool* fq_threadpool)
-    : fq_threadpool_{fq_threadpool} {
-}
-
-io::Result<std::pair<io::Sink*, uint8_t>, GenericError> FileSnapshotStorage::OpenFile(
-    const std::string& path) {
-  if (fq_threadpool_) {  // EPOLL
-    auto res = util::OpenFiberWriteFile(path, fq_threadpool_);
-    if (!res) {
-      return nonstd::make_unexpected(GenericError(res.error(), "Couldn't open file for writing"));
-    }
-
-    return std::pair(*res, FileType::FILE);
-  } else {
-#ifdef __linux__
-    auto res = OpenLinux(path, kRdbWriteFlags, 0666);
-    if (!res) {
-      return nonstd::make_unexpected(GenericError(
-          res.error(),
-          "Couldn't open file for writing (is direct I/O supported by the file system?)"));
-    }
-
-    uint8_t file_type = FileType::FILE | FileType::IO_URING;
-    if (kRdbWriteFlags & O_DIRECT) {
-      file_type |= FileType::DIRECT;
-    }
-    return std::pair(new LinuxWriteWrapper(res->release()), file_type);
-#else
-    LOG(FATAL) << "Linux I/O is not supported on this platform";
-#endif
-  }
-}
-
-AwsS3SnapshotStorage::AwsS3SnapshotStorage(util::cloud::AWS* aws) : aws_{aws} {
-}
-
-io::Result<std::pair<io::Sink*, uint8_t>, GenericError> AwsS3SnapshotStorage::OpenFile(
-    const std::string& path) {
-  DCHECK(aws_);
-
-  optional<pair<string, string>> bucket_path = GetBucketPath(path);
-  if (!bucket_path) {
-    return nonstd::make_unexpected(GenericError("Invalid S3 path"));
-  }
-  auto [bucket_name, obj_path] = *bucket_path;
-
-  cloud::S3Bucket bucket(*aws_, bucket_name);
-  error_code ec = bucket.Connect(kBucketConnectMs);
-  if (ec) {
-    return nonstd::make_unexpected(GenericError(ec, "Couldn't connect to S3 bucket"));
-  }
-  auto res = bucket.OpenWriteFile(obj_path);
-  if (!res) {
-    return nonstd::make_unexpected(GenericError(res.error(), "Couldn't open file for writing"));
-  }
-
-  return std::pair<io::Sink*, uint8_t>(*res, FileType::CLOUD);
 }
 
 string InferLoadFile(string_view dir, cloud::AWS* aws) {
