+++ src/server/detail/save_stages_controller.h
@@ -7,6 +7,7 @@
 
 #include <filesystem>
 
+#include "server/detail/snapshot_storage.h"
 #include "server/rdb_save.h"
 #include "server/server_family.h"
 #include "util/cloud/aws.h"
@@ -18,45 +19,6 @@ class Transaction;
 class Service;
 
 namespace detail {
-
-enum FileType : uint8_t {
-  FILE = (1u << 0),
-  CLOUD = (1u << 1),
-  IO_URING = (1u << 2),
-  DIRECT = (1u << 3),
-};
-
-class SnapshotStorage {
- public:
-  virtual ~SnapshotStorage() = default;
-
-  // Opens the file at the given path, and returns the open file and file
-  // type, which is a bitmask of FileType.
-  virtual io::Result<std::pair<io::Sink*, uint8_t>, GenericError> OpenFile(
-      const std::string& path) = 0;
-};
-
-class FileSnapshotStorage : public SnapshotStorage {
- public:
-  FileSnapshotStorage(FiberQueueThreadPool* fq_threadpool);
-
-  io::Result<std::pair<io::Sink*, uint8_t>, GenericError> OpenFile(
-      const std::string& path) override;
-
- private:
-  util::fb2::FiberQueueThreadPool* fq_threadpool_;
-};
-
-class AwsS3SnapshotStorage : public SnapshotStorage {
- public:
-  AwsS3SnapshotStorage(util::cloud::AWS* aws);
-
-  io::Result<std::pair<io::Sink*, uint8_t>, GenericError> OpenFile(
-      const std::string& path) override;
-
- private:
-  util::cloud::AWS* aws_;
-};
 
 struct SaveStagesInputs {
   bool use_dfs_format_;
